+++ aim/web/ui/src/pages/ImagesExplore/ImagesExplore.tsx
@@ -194,6 +194,7 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
               applyButtonDisabled={imagesExploreData?.applyButtonDisabled}
               panelResizing={panelResizing}
               resizeMode={imagesExploreData?.config?.table.resizeMode}
+              tableHeight={imagesExploreData?.config?.table?.height}
               imageWrapperOffsetHeight={offsetHeight || 0}
               imageWrapperOffsetWidth={offsetWidth || 0}
               isRangePanelShow={
@@ -208,6 +209,9 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
               tooltip={
                 imagesExploreData?.config?.images?.tooltip as IPanelTooltip
               }
+              imageProperties={
+                imagesExploreData?.config?.images?.imageProperties
+              }
               onActivePointChange={imagesExploreAppModel.onActivePointChange}
               controls={
                 <Controls
@@ -217,7 +221,17 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
                   tooltip={
                     imagesExploreData?.config?.images?.tooltip as IPanelTooltip
                   }
+                  imageProperties={
+                    imagesExploreData?.config?.images?.imageProperties
+                  }
                   onChangeTooltip={imagesExploreAppModel?.onChangeTooltip}
+                  onImageSizeChange={imagesExploreAppModel.onImageSizeChange}
+                  onImageRenderingChange={
+                    imagesExploreAppModel.onImageRenderingChange
+                  }
+                  onImageAlignmentChange={
+                    imagesExploreAppModel.onImageAlignmentChange
+                  }
                 />
               }
             />