+++ aim/web/ui/src/services/models/imagesExplore/imagesExploreAppModel.ts
@@ -6,7 +6,14 @@ import { saveAs } from 'file-saver';
 import { RowHeightSize } from 'config/table/tableConfigs';
 import { BookmarkNotificationsEnum } from 'config/notification-messages/notificationMessages';
 import { ResizeModeEnum, RowHeightEnum } from 'config/enums/tableEnums';
-import { blobsUpdateThrottleDelay } from 'config/imagesConfigs/imagesConfig';
+import {
+  blobsUpdateThrottleDelay,
+  IMAGE_SIZE_CHANGE_DELAY,
+} from 'config/imagesConfigs/imagesConfig';
+import {
+  ImageAlignmentEnum,
+  ImageRenderingEnum,
+} from 'config/enums/imageEnums';
 
 import {
   getImagesExploreTableColumns,
@@ -96,6 +103,11 @@ function getConfig(): IImagesExploreAppConfig {
         content: {},
         display: true,
         selectedParams: [],
+      },
+      imageProperties: {
+        alignmentType: ImageAlignmentEnum.Height,
+        imageSize: 15,
+        imageRendering: ImageRenderingEnum.Pixelated,
       },
       focusedState: {
         active: false,
@@ -460,6 +472,7 @@ function setModelData(rawData: any[], configData: IImagesExploreAppConfig) {
       active: false,
       key: null,
     },
+    imageProperties: config.images.imageProperties,
   };
   model.setState({
     requestIsPending: false,
@@ -1837,6 +1850,75 @@ function onRecordDensityChange(event: ChangeEvent<{ value: number }>) {
   }
 }
 
+const onImageSizeChange = _.throttle((value: number) => {
+  const configData: IImagesExploreAppConfig | undefined =
+    model.getState()?.config;
+  if (configData?.images) {
+    const images = {
+      ...configData.images,
+      imageProperties: {
+        ...configData.images.imageProperties,
+        imageSize: value,
+      },
+    };
+    const config = {
+      ...configData,
+      images,
+    };
+    updateURL(config as IImagesExploreAppConfig);
+    model.setState({
+      config,
+    });
+  }
+}, IMAGE_SIZE_CHANGE_DELAY);
+
+function onImageRenderingChange(type: ImageRenderingEnum) {
+  const configData: IImagesExploreAppConfig | undefined =
+    model.getState()?.config;
+  if (configData?.images) {
+    const images = {
+      ...configData.images,
+      imageProperties: {
+        ...configData.images.imageProperties,
+
+        imageRendering: type,
+      },
+    };
+    const config = {
+      ...configData,
+      images,
+    };
+    updateURL(config as IImagesExploreAppConfig);
+    model.setState({
+      config,
+    });
+  }
+}
+
+function onImageAlignmentChange(
+  option: { value: string; label: string } | null,
+) {
+  const configData: IImagesExploreAppConfig | undefined =
+    model.getState()?.config;
+  if (configData?.images) {
+    const images = {
+      ...configData.images,
+      imageProperties: {
+        ...configData.images.imageProperties,
+        alignmentType: option?.value,
+      },
+    };
+    const config = {
+      ...configData,
+      images,
+    };
+    updateURL(config as IImagesExploreAppConfig);
+    model.setState({
+      config,
+    });
+  }
+}
+
 const imagesExploreAppModel = {
   ...model,
   initialize,
@@ -1879,6 +1961,9 @@ const imagesExploreAppModel = {
   getImagesBlobsData,
   onChangeTooltip,
   onActivePointChange,
+  onImageSizeChange,
+  onImageRenderingChange,
+  onImageAlignmentChange,
 };
 
 export default imagesExploreAppModel;